# api-docs.json
{"swagger":"2.0","info":{"description":"An application for persons to find events created by ongs . UserToEvent entity does not appear in swagger documentation","version":"Version 1","title":"Proiect Programare Web Java","termsOfService":"Terms of service","contact":{"name":"sichitiu.marian@yahoo.com"},"license":{"name":"License of API","url":"https://swagger.io/docs/"}},"host":"localhost:8090","basePath":"/","tags":[{"name":"admin-controller","description":"Admin Controller"},{"name":"change-pass-controller","description":"Change Pass Controller"},{"name":"index-controller","description":"Index Controller"},{"name":"new-users-controller","description":"New Users Controller"},{"name":"ong-controller","description":"Ong Controller"},{"name":"person-controller","description":"Person Controller"}],"paths":{"/":{"get":{"tags":["index-controller"],"summary":"Get last five elements","operationId":"getEventImagesUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Event"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/admin/acceptong/{ongId}":{"post":{"tags":["admin-controller"],"summary":"Admin accepts ong","operationId":"acceptOngPostUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"ongId","in":"path","description":"The id of the ong user","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/admin/deleteOng/{ongId}":{"delete":{"tags":["admin-controller"],"summary":"Delete ong user by given id","operationId":"deleteOngUsingDELETE","produces":["*/*"],"parameters":[{"name":"ongId","in":"path","description":"The id of the ong user","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Ong"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/admin/deletePerson/{personId}":{"delete":{"tags":["admin-controller"],"summary":"Delete person user by given id","operationId":"deletePersonUsingDELETE","produces":["*/*"],"parameters":[{"name":"personId","in":"path","description":"The id of the person user","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Person"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/admin/ongs":{"get":{"tags":["admin-controller"],"summary":"Get all ong users as a list.","operationId":"ongsUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Ong"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/admin/persons":{"get":{"tags":["admin-controller"],"summary":"Get all person users as a list.","operationId":"personsUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Person"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/changepassword/{username}":{"post":{"tags":["change-pass-controller"],"summary":"Change user password","operationId":"changePasswordUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"parameters","in":"query","description":"Password parameters","required":true,"items":{"type":"object","additionalProperties":{"type":"string"}}},{"name":"username","in":"path","description":"The username of the user","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/User"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/createong":{"put":{"tags":["new-users-controller"],"summary":"Create Ong user","operationId":"createOngUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"ong","description":"The ong parameters","required":true,"schema":{"$ref":"#/definitions/Ong"}}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/createperson":{"put":{"tags":["new-users-controller"],"summary":"Create Person user","operationId":"createPersonUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"person","description":"The person parameters","required":true,"schema":{"$ref":"#/definitions/Person"}}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/event/delete/{eventid}":{"delete":{"tags":["ong-controller"],"summary":"Delete event created by ong","operationId":"deleteEventUsingDELETE","produces":["*/*"],"parameters":[{"name":"eventid","in":"path","description":"The id of the event","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/login":{"get":{"tags":["index-controller"],"summary":"Get last five elements","operationId":"getEventImagesUsingGET_1","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Event"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["index-controller"],"summary":"Login user with username and password","operationId":"loginUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"parameters","in":"query","description":"The username and password of the user","required":true,"items":{"type":"object","additionalProperties":{"type":"string"}}}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/logout":{"get":{"tags":["index-controller"],"summary":"Logout and Destroy session parameter","operationId":"logoutUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/ong/event":{"post":{"tags":["ong-controller"],"summary":"Create event by ong with given fields","operationId":"saveEventUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"parameters","in":"query","description":"The parameters of the event","required":true,"items":{"type":"object","additionalProperties":{"type":"string"}}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Event"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/ong/event/update":{"post":{"tags":["ong-controller"],"summary":"Update event by ong with given fields","operationId":"updateUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"event","description":"The parameters of the event","required":true,"schema":{"$ref":"#/definitions/Event"}}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/ong/myEvents":{"get":{"tags":["ong-controller"],"summary":"Display Events by the logged in ong","operationId":"getEventsUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Event"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/person/allevents":{"get":{"tags":["person-controller"],"summary":"Show events joined by person logged in","operationId":"alleventsUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Event"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/person/event/delete/{eventid}":{"delete":{"tags":["person-controller"],"summary":"Remove event joined by person","operationId":"deleteUser2eventUsingDELETE","produces":["*/*"],"parameters":[{"name":"eventid","in":"path","description":"The event id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/person/joinevent/{eventid}":{"put":{"tags":["person-controller"],"summary":"Person user joins event","operationId":"addEventUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"eventid","in":"path","description":"The event of the id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Event"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/person/profile":{"get":{"tags":["person-controller"],"summary":"Show logged in person profile","operationId":"profileUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Person"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/updateOng":{"post":{"tags":["ong-controller"],"summary":"Update ong fields","operationId":"updateONGUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"ong","description":"The ong parameters","required":true,"schema":{"$ref":"#/definitions/Ong"}}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/updatePerson":{"post":{"tags":["person-controller"],"summary":"Update person fields","operationId":"updatePersonUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"person","description":"The person parameters","required":true,"schema":{"$ref":"#/definitions/Person"}}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}}},"definitions":{"Event":{"type":"object","properties":{"approved":{"type":"boolean"},"date":{"type":"string","format":"date-time"},"description":{"type":"string"},"id":{"type":"integer","format":"int32"},"location":{"$ref":"#/definitions/Location"},"name":{"type":"string"},"ong":{"$ref":"#/definitions/Ong"}},"title":"Event"},"Location":{"type":"object","properties":{"address":{"type":"string"},"location_id":{"type":"integer","format":"int32"},"postalCode":{"type":"string"}},"title":"Location"},"Ong":{"type":"object","properties":{"approved":{"type":"boolean"},"contact":{"type":"string"},"email":{"type":"string"},"fiscalcode":{"type":"string","pattern":"\\d+"},"fullName":{"type":"string"},"id":{"type":"integer","format":"int32"},"isadmin":{"type":"boolean"},"password":{"type":"string"},"phone":{"type":"string","pattern":"\\d+"},"username":{"type":"string"}},"title":"Ong"},"Person":{"type":"object","properties":{"age":{"type":"integer","format":"int32","minimum":15,"maximum":90,"exclusiveMinimum":false,"exclusiveMaximum":false},"email":{"type":"string"},"fullName":{"type":"string"},"id":{"type":"integer","format":"int32"},"isadmin":{"type":"boolean"},"job":{"type":"string"},"password":{"type":"string"},"phone":{"type":"string","pattern":"\\d+"},"picture":{"type":"string"},"username":{"type":"string"}},"title":"Person"},"User":{"type":"object","properties":{"email":{"type":"string"},"id":{"type":"integer","format":"int32"},"isadmin":{"type":"boolean"},"password":{"type":"string"},"username":{"type":"string"}},"title":"User"}}}